2020/11/10 memory leak in mod function

+++ BasicOperation.c
@@ -67,15 +67,15 @@ unsigned int hex2int(char c)
 {
 	if (c >= 'a' && c <= 'z')
 	{
-		return c - 'a' + 10;
+		return (int)c - 'a' + 10;
 	}
 	else if (c >= 'A' && c <= 'Z')
 	{
-		return c - 'A' + 10;
+		return (int)c - 'A' + 10;
 	}
 	else if (c >= '0' && c <= '9')
 	{
-		return c - '0';
+		return (int)c - '0';
 	}
 	else
 	{
@@ -161,7 +161,6 @@ ErrorMessage wordSquaring(word* C1, word* C0, word A)
 
 	return SUCCESS;
 }
-//TODO
 ErrorMessage wordLongDivision(word* Q, word A1, word A0, word B)
 {
 	*Q = 0;
